implement an interface to support DNS challenges
fixes #51

As the ability to install DNS challenges into the infrastructure depends
on the site, I implemented this in a way that allows cfookbook authores
to specify how to supply this to the infrastructure using custom ruby
blocks.

There must be two blocks given to the resource if you want to use DNS
validation:

`install_authz_block` and `remove_authz_block`

There is an example in the README explaining how to use this.